requirements.txt
ccxt
python-dotenv
PyYAML
pandas
numpy
aiohttp
